Move signed-in audit board page component into own module.
